fix: rename directory
