+++ jest/sfccPathSetup.js
@@ -401,9 +401,9 @@ jest.mock(
 );
 
 jest.mock(
-  '*/cartridge/client/default/js/adyen_checkout/adyenGiftcards',
+  '*/cartridge/client/default/js/adyen_checkout/giftcards',
   () =>
-    require('../src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/adyenGiftcards'),
+    require('../src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/giftcards'),
   { virtual: true },
 );
