+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/__tests__/renderGiftcardComponent.test.js
@@ -6,7 +6,7 @@ const {
   renderAddedGiftCard,
   renderGiftCardSelectForm,
   showGiftCardWarningMessage,
-} = require('../adyenGiftcards');
+} = require('../giftcards');
 const store = require('../../../../../store');
 
 jest.mock('../../commons');
