+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/giftcards/index.js
