+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/paymentMethodsConfiguration/giftcards/giftcardsConfig.js
@@ -5,7 +5,7 @@ const {
   showGiftCardCancelButton,
   attachGiftCardCancelListener,
   createElementsToShowRemainingGiftCardAmount,
-} = require('../../adyenGiftcards');
+} = require('../../giftcards');
 const { GIFTCARD, SUCCESS, NOTENOUGHBALANCE } = require('../../../constants');
 const { initializeCheckout } = require('../../renderGenericComponent');
