+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/renderGenericComponent.js
@@ -19,7 +19,7 @@ const {
   clearGiftCardsContainer,
   attachGiftCardCancelListener,
   showGiftCardCancelButton,
-} = require('./adyenGiftcards');
+} = require('./giftcards');
 
 function addPosTerminals(terminals) {
   const ddTerminals = document.createElement('select');
